Only add newline after hard line break if needed, fixes #80
If `preserveWhitespace: 'full'` is set, hard line breaks are
parsed as `<br \>\n` so the newline character after the tag is not removed.
In this case there must be no additional newline character added when
serializing the `<br \>` tag to markdown as this would create a new block and break marks.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>